added createZone route
